+++ software/audioctl_ui/index.html
@@ -10,6 +10,6 @@
   </head>
   <body>
     <div id="mixer"></div>
-    <script type="module" src="/src/main.ts"></script>
+    <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
