+++ software/audioctl_ui/package.json
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "misirka": "github:dexterlb/misirka#97edb75",
+    "preact": "^10.27.2",
     "zod": "^4.3.6"
   },
   "devDependencies": {
